pyproject.toml
[tool.poetry]
name = "byodb"
version = "0.1.0"
description = "A personal database that powers privacy-respecting apps."
authors = ["Beto Dealmeida <roberto@dealmeida.net>"]
readme = "README.md"
packages = [{include = "byodb", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
quart = "^0.19.3"
quart-schema = "^0.18.0"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8-pyproject = "^1.2.3"
pytest-asyncio = "^0.23.2"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
freezegun = "^1.3.1"
pre-commit = "^3.6.0"
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "byodb.main:run"
init_db = "byodb.main:init_db_sync"

[tool.flake8]
max-line-length = 90

[tool.pytest.ini_options]
asyncio_mode = "auto"
